Remove ui and pageable
