+++ INSTALL.md
@@ -7,7 +7,7 @@ Install python 2.x (will probably work with 3.x with little modification)
 Install requirements:
 
     sudo apt install python-pyside python-pip python-dev
-    pip install -r requirements.txt
+    pip install --user -r requirements.txt
 
 Configure the library_dir in [ui/config.rc](ui/config.rc) to point at the [books](books) directory which is full of test books.
 
@@ -19,7 +19,7 @@ Run the tests:
 Run the UI using the emulator:
 
     cd ui/
-    python ui.py
+    python main.py
 
 ## Raspberry Pi
 
@@ -35,5 +35,5 @@ Copy the test books to the home directory:
 Run the UI software without the emulator:
 
     cd ui/
-    python ui.py --disbale-emulator
+    python main.py --disbale-emulator
